move unicodedata to internal/unicodedata
