+++ internal/unicodedata/east_asian_width.go
