+++ internal/unicodedata/emojis.go
