+++ internal/unicodedata/general_category.go
