+++ internal/unicodedata/linebreak.go
