+++ internal/unicodedata/vertical_orientation.go
