+++ segmenter/segmenter.go
@@ -15,7 +15,7 @@ package segmenter
 import (
 	"unicode"
 
-	ucd "github.com/go-text/typesetting/unicodedata"
+	ucd "github.com/go-text/typesetting/internal/unicodedata"
 )
 
 // breakAttr is a flag storing the break properties between two runes of
